src/test/javascript/spec/app/entities/list-services/list-services-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';

import AlertService from '@/shared/alert/alert.service';
import * as config from '@/shared/config/config';
import ListServicesUpdateComponent from '@/entities/list-services/list-services-update.vue';
import ListServicesClass from '@/entities/list-services/list-services-update.component';
import ListServicesService from '@/entities/list-services/list-services.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.component('font-awesome-icon', {});

describe('Component Tests', () => {
  describe('ListServices Management Update Component', () => {
    let wrapper: Wrapper<ListServicesClass>;
    let comp: ListServicesClass;
    let listServicesServiceStub: SinonStubbedInstance<ListServicesService>;

    beforeEach(() => {
      listServicesServiceStub = sinon.createStubInstance<ListServicesService>(ListServicesService);

      wrapper = shallowMount<ListServicesClass>(ListServicesUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          alertService: () => new AlertService(store),
          listServicesService: () => listServicesServiceStub
        }
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.listServices = entity;
        listServicesServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(listServicesServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.listServices = entity;
        listServicesServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(listServicesServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });
  });
});
